pyproject.toml
[tool.poetry]
name = "neural-networks"
version = "0.1.2"
description = "Neural Network Toolkit for Speech and Language Processing"
authors = ["Kosuke Mori"]
homepage = "https://github.com/kmori6"
repository = "https://github.com/kmori6/neural-networks"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
torchaudio = "^2.3.0"
sentencepiece = "^0.2.0"
openai-whisper = "^20231117"
hydra-core = "^1.3.2"
soundfile = "^0.12.1"
librosa = "^0.10.1"
sacrebleu = "^2.4.2"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[tool.poetry.group.test.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "./"
